samples/websockets: Restructure WebSocket handler as an async event loop
Implement an event loop based on kqueue and coroutines.  This allows
reasonable concurrency by blocking until an event occurs and resuming a
coroutine for the event, which can then yield back to the event loop
when it is ready to wait for the next event.

For now the only event is the websocket having data ready to read.  Soon
there will be pipes to read from long running child processes, and this
enables websocket messages to continue to be handled without busy-loop
polling descriptors for data.