ws: Serve scripts/*.js as static content
Use sendfile(3lua) from ryan-moeller/flualibs to serve .js files in the
scripts/ directory as static content.  This is more an example of the
concept than a practical improvement at this stage, as the one script is
pretty small.  It would be much more practical for serving large files
such as packages or disk images or media.