samples/websockets: Add helper to read from a child pipe
This function creates a coroutine that is resumed whenever there is
data to read on a pipe and invokes a callback with the accumulated data
every time a newline is encountered.  When the pipe is closed the
coroutine waits for a pid to exit, optionally invokes an exit callback,
and finally removes itself from the thread registry.