httpd: File header cleanup
Use SPDX-License-Identifier in place of the full text of the license.

Remove a comment about using with inetd, which is misleading.  A server
utilizing this module could be used with inetd, not the module itself.

Move the modeline to the bottom of the file, and while here add some
extra blank lines around a top level definition for consistency.