Handle chunk-encoded body transfers
Recognize a "Transfer-Encoding: chunked" header and set the body of the
request to a function that returns an iterator over the chunks.

Trailers following the body are returned in the trailers field of the
request after the iterator has consumed the body.

The handler should consume the full body before returning, though in
practice it should not currently matter as pipelined requests are not
yet supported.