httpd: Avoid exiting after the response if possible
When the server input and output are standard IO streams, we cannot
close the files, so are forced to exit the process.  This works for the
current architecture where the server is invoked by inetd.  But for
testing or other potential server architectures, we should avoid exiting
if possible.

Check whether the input is stdin or output is stdout, and if not, close
the server's file handles instead of exiting the process.