httpd: Shortcut NOP execution
Avoid testing bits one by one for NOP.  It is the most common opcode,
filling 231 out of 271 entries of the header value parser LUTs.  No
state change tends to mean no operation to perform.

We can easily check for NOP and shortcut the opcode decoding.

While here, comment out the line stashing the opcode in the parser.
This was left over from an eliminated feature that used the free
least-significant bit of the opcode as an offset for MARK.  Now it is
only useful for debugging.