httpd: Allow sending a list of headers with the same name
Sending multiple field-lines with the same field-name is permitted, so
enable this by checking if the value is a table and iterating as a list
if so.

While here, add comments justifying the implementation of headers and
cookies as tables with citations for the relevant RFCs.

Also document that cookie values are expected to encode any attributes.