README.md: Describe alternatives to inetd
For some purposes, superserver-style socket activation is not ideal.

`socat` and `ncat` provide an inetd-style single-listener frontend,
and can even handle TLS for HTTPS.

Add several examples of `socat` and `ncat` one-liners.

While `inetd` is included with typical BSD operating systems, it is not
standard in macOS or most Linux distributions.

Add rough examples for `launchd` and `systemd`.