httpd: Handle boundless request body fallback
In HTTP/1.1, a message may omit Content-Length and forego chunked
transfer encoding, in which case the message is ended by closing the
conection.

Handle messages delimited by connection close when receiving a request.

This wraps the body in an iterator over buffered pieces, similar to how
chunk-transferred bodies are exposed.