httpd: Provide normalized field name access
Request fields are already normalized.  Add helpers for normalized
access to response fields.  This simplifies code that has to manipulate
response fields.

While here, normalize the terms used to "field" rather than "header",
since there has been a tendency to use "header" for identifiers in code
that is really agnostic to whether it is dealing with header fields or
trailer fields.  This is also nice because "field" is shorter.