httpd: Implement Bad Request responses
Restructure request handling to enable writing error responses during
body handling.

Alter the server state machine to enable checking whether we have
performed an error response during a handler.

Improve conformance to RFC 9112 §6.1 ("Transfer-Encoding") by checking
that "chunked" is the last coding listed in transfer-encoding and §6.3
("Message Body Length") by checking for response statuses that are not
allowed a body, regardless of headers, and skipping the body.  The one
exception is for Switching Protocols, where the body function takes over
for the new protocol (so it's technically not a response body at that
point).