httpd: Improve logger detail and control
Add a parameter to `create_server` to override the server identifier,
which is the server PID obtained from `$PPID` of an `io.popen` shell by
default.

Add a third loop variable to be returned by a listener accept iterator.
The new value is a string label for the client to be used in log
messages.  Logs outside of `server:accept()` handling are labeled
`"(server)"`, and the `StdioListener` returns the label `"(stdio)"`.

Add a third parameter to `server:accept()` for the label, with the
default `"(client)"`.