samples: Take advantage of newly-optional kevent fields
As flualibs/kqueue has been updated to make some kevent fields optional,
take advantage of this to explicitly manage event and thread lifetimes
directly in the event loop.

We can now check if an event has an associated thread.  If so, we try to
resume the thread, logging any errors.  If the thread is dead after
being resumed, we delete the corresponding event from the queue then can
safely finalize the thread.

While here, log event errors for events with a thread.  If the event has
no thread, it was probably due to the deletion of the event failing with
ENOENT for an event that had already removed itself, which does not need
to be logged.